backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
alembic==1.12.1 
python-dotenv==1.0.1
boto3==1.34.0
botocore==1.34.0
celery==5.3.4
redis==5.0.1